Ensure terraform plan output is logged
When using `returnStdout` with the `sh()` method, the command's output
does not show up in the job log. We want both to happen - to capture the
output and ensure it shows up in the job log. To remedy that, store the
output to a variable, echo it out, and return the value.