Implement terraform fmt
This commit introduces a new `terraform.fmt` method to check for proper
formatting of Terraform code (see issue #5 for more information on the
feature). The use-case is for validating that Terraform code
contributions have been properly formatted, and erroring out should that
not be the case.  Exposed parameters are documented (including the
mutual exclusivity between `check` and `write`), and the test
`Jenkinsfile` has been updated with a new stage for `fmt`.